third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_multiple_pointerover_no_pointer_movement-expected.txt

This is a testharness.js-based test.
[FAIL] Pointer boundary events should be fired per pointer
  assert_equals: expected "pointerover@div#target, pointerover@div#target, pointerover@div#container, pointerover@div#container" but got "pointerover@div#target, pointerover@div#container"
Harness: the test ran to completion.
